backend/requirements.txt
# Core Flask Stack
flask==3.1.0
werkzeug==3.1.3
flask-cors
gunicorn
python-dotenv

# AI and ML Libraries
torch>=2.0.0
transformers
diffusers
accelerate
replicate
openai
anthropic
google-generativeai

# Image Processing
pillow
opencv-python-headless
imagehash
scikit-image

# Data Manipulation
numpy
scipy
scikit-learn
pandas

# Visualization
matplotlib
seaborn
plotly

# Video Processing
moviepy
pytube

# Gradio
gradio

# Logging and Utilities
requests

# Machine Learning
ultralytics  # Latest version now that we're using Python 3.10+

# Mochi Fine-tuning Dependencies
#git+https://github.com/a-r-r-o-w/cogvideox-factory.git
torchvision
torchaudio
einops
wandb
lightning
hydra-core
